--- types/.eslintrc
@@ -1,4 +1,0 @@
-{
-  "root": true,
-  "parser": "babel-eslint"
-}
