+++ types/ClassNames.d.ts
@@ -1,0 +1,33 @@
+export interface ClassNames {
+  container: string;
+  wrapper: string;
+  interactionDisabled: string;
+  navBar: string;
+  navButtonPrev: string;
+  navButtonNext: string;
+  navButtonInteractionDisabled: string;
+
+  months: string;
+  month: string;
+  caption: string;
+  weekdays: string;
+  weekdaysRow: string;
+  weekday: string;
+  weekNumber: string;
+  body: string;
+  week: string;
+  day: string;
+  footer: string;
+  todayButton: string;
+
+  today: string;
+  selected: string;
+  disabled: string;
+  outside: string;
+}
+
+export interface InputClassNames {
+  container: string;
+  overlayWrapper: string;
+  overlay: string;
+}
