+++ types/DayPickerInput.d.ts
@@ -1,8 +1,11 @@
 import * as React from 'react';
-import { DayPickerInputProps } from './props';
+import { DayPickerInputProps } from './Props';
 import DayPicker from './DayPicker';
 
-export class DayPickerInput extends React.Component<DayPickerInputProps, any> {
+export default class DayPickerInput extends React.Component<
+  DayPickerInputProps,
+  any
+> {
   showDayPicker: () => void;
   hideDayPicker: () => void;
   getDayPicker: () => DayPicker;
