+++ types/Modifiers.d.ts
@@ -1,55 +1,32 @@
-export interface ClassNames {
-  container: string;
-  wrapper: string;
-  interactionDisabled: string;
-  navBar: string;
-  navButtonPrev: string;
-  navButtonNext: string;
-  navButtonInteractionDisabled: string;
-
-  months: string;
-  month: string;
-  caption: string;
-  weekdays: string;
-  weekdaysRow: string;
-  weekday: string;
-  weekNumber: string;
-  body: string;
-  week: string;
-  day: string;
-  footer: string;
-  todayButton: string;
-
-  today: string;
-  selected: string;
-  disabled: string;
-  outside: string;
-}
-
-export interface InputClassNames {
-  container: string;
-  overlayWrapper: string;
-  overlay: string;
+export interface RangeModifier {
+  from: Date;
+  to: Date;
 }
 
 export interface RangeModifier {
   from: Date;
   to: Date;
 }
+
 export interface BeforeModifier {
   before: Date;
 }
+
 export interface AfterModifier {
   after: Date;
 }
+
 export interface BeforeAfterModifier {
   after: Date;
   before: Date;
 }
+
 export interface DaysOfWeekModifier {
   daysOfWeek: number[];
 }
+
 export type FunctionModifier = (date: Date) => boolean;
+
 export type Modifier =
   | Date
   | RangeModifier
@@ -71,3 +48,12 @@ export interface DayModifiers {
   outside: boolean | undefined;
   [other: string]: boolean | undefined;
 }
+
+export const ModifiersUtils: {
+  dayMatchesModifier(day: Date, modifier?: Modifier | Modifier[]): boolean;
+  getModifiersForDay(
+    day: Date,
+    modifiers: Record<string, Modifier | Modifier[]>
+  ): string[];
+};
+export type ModifiersUtils = typeof ModifiersUtils;
