+++ types/MomentLocaleUtils.d.ts
@@ -1,2 +1,2 @@
-import { LocaleUtils } from './utils';
+import { LocaleUtils } from './LocaleUtils';
 export const MomentLocaleUtils: LocaleUtils;
