+++ types/tsconfig.json
@@ -1,13 +1,14 @@
 {
   "compilerOptions": {
     "module": "commonjs",
-    "lib": ["es6", "dom"],
+    "lib": ["es6"],
     "noImplicitAny": true,
     "noImplicitThis": true,
+    "strictFunctionTypes": true,
     "strictNullChecks": true,
+    "types": [],
     "noEmit": true,
-    "forceConsistentCasingInFileNames": true,
-    "strictFunctionTypes": false
+    "forceConsistentCasingInFileNames": true
   },
-  "include": ["./*.d.ts"]
+  "files": ["index.d.ts"]
 }
